[tests] Separate the Git interceptor
Move the shim protocol, typed Git faults, and raw push recording out
of the GraphQL fake.

Limit command recognition to GHerrit's stable subcommand-first shape
and keep every Git argument opaque.

gherrit-pr-id: Gjycdin4ykmksmeu7nohrc5y6f3qrq4h4
