+++ testutil/src/lib.rs
@@ -19,6 +19,7 @@ use regex::Regex;
 use tempfile::TempDir;
 
 mod command;
+mod git_interceptor;
 mod mock_server;
 
 pub use command::TestCommand;
@@ -876,9 +877,13 @@ impl TestContext {
         assert!(self.has_git_interceptor, "missing test capability: .with_git_interceptor()");
         self.inspect_mock_state(|state| {
             state
-                .pushes
+                .git
+                .pushes()
                 .iter()
-                .map(|push| PushRecord { arguments: push.args.clone(), exit_code: push.exit_code })
+                .map(|push| PushRecord {
+                    arguments: push.arguments().to_vec(),
+                    exit_code: push.exit_code(),
+                })
                 .collect()
         })
     }
